v1.0.0 ending

+++ src/main/java/com/admin/web/service/SysMenuService.java
@@ -1,9 +1,13 @@
 package com.admin.web.service;
 
 import com.admin.web.dao.SysMenuDao;
+import com.admin.web.exception.WebServerException;
+import com.admin.web.model.ServerResponseEntity;
 import com.admin.web.model.SysMenu;
 import com.admin.web.model.SysUser;
 import com.admin.web.model.enums.Move;
+import com.admin.web.model.vo.MoveVo;
+import com.admin.web.utils.BeanUtils;
 import com.admin.web.utils.SecurityUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,11 +25,11 @@ public class SysMenuService {
         this.sysMenuDao = sysMenuDao;
     }
 
-    public List<SysMenu> findAll() {
+    public List<SysMenu> all() {
         return this.sysMenuDao.findByOrderBySort();
     }
 
-    public List<SysMenu> findAll(SysUser sysUser) {
+    public List<SysMenu> all(SysUser sysUser) {
         if (SecurityUtils.isSuperAdmin(sysUser)) {
             //超级管理员可以看到所有菜单
             return this.sysMenuDao.findBySysMenuOrderBySort(true);
@@ -38,27 +42,10 @@ public class SysMenuService {
         return this.sysMenuDao.findByUserIdAndDisableOrderBySort(sysUser.getId(), false);
     }
 
-    public SysMenu findByTitle(String title) {
-        return this.sysMenuDao.findByTitle(title);
-    }
-
-    public Optional<SysMenu> findById(Long id) {
-        return this.sysMenuDao.findById(id);
-    }
-
     @Transactional(rollbackFor = Exception.class)
-    public void save(SysMenu sysMenu) {
-        if (Objects.nonNull(sysMenu.getId())) {
-            this.sysMenuDao.save(sysMenu);
-        } else {
-            this.sysMenuDao.save(sysMenu);
-            sysMenu.setSort(sysMenu.getId());
-            this.sysMenuDao.save(sysMenu);
-        }
-    }
-
-    @Transactional(rollbackFor = Exception.class)
-    public void move(SysMenu sysMenu, Move move) {
+    public void move(MoveVo vo) {
+        SysMenu sysMenu = this.sysMenuDao.findById(vo.getId())
+                .orElseThrow(() -> new WebServerException(ServerResponseEntity.fail("菜单不存在！")));
         List<SysMenu> sysMenus = this.sysMenuDao.findByPidOrderBySort(sysMenu.getPid());
         int index = -1;
         for (int i = 0; i < sysMenus.size(); i++) {
@@ -67,8 +54,8 @@ public class SysMenuService {
                 break;
             }
         }
-        if (Move.UP == move ? index > 0 : index < sysMenus.size() - 1 && index != -1) {
-            SysMenu oldSysMenu = sysMenus.get(Move.UP == move ? index - 1 : index + 1);
+        if (Move.UP == vo.getMove() ? index > 0 : index < sysMenus.size() - 1 && index != -1) {
+            SysMenu oldSysMenu = sysMenus.get(Move.UP == vo.getMove() ? index - 1 : index + 1);
             Long oldSort = oldSysMenu.getSort();
             oldSysMenu.setSort(sysMenu.getSort());
             sysMenu.setSort(oldSort);
@@ -77,23 +64,56 @@ public class SysMenuService {
         }
     }
 
-    public void deleteById(Long id) {
+    @Transactional(rollbackFor = Exception.class)
+    public void create(SysMenu sysMenu) {
+        if (Objects.nonNull(sysMenu.getPid())
+                && !this.sysMenuDao.existsById(sysMenu.getPid())) {
+            throw new WebServerException(ServerResponseEntity.fail("上级菜单不存在！"));
+        }
+        if (Objects.nonNull(this.sysMenuDao.findByTitle(sysMenu.getTitle()))) {
+            throw new WebServerException(ServerResponseEntity.fail("菜单标题已存在！"));
+        }
+        this.sysMenuDao.save(sysMenu);
+        sysMenu.setSort(sysMenu.getId());
+        this.sysMenuDao.save(sysMenu);
+    }
+
+    public void update(SysMenu sysMenu) {
+        SysMenu oldSysMenu = this.sysMenuDao.findById(sysMenu.getId())
+                .orElseThrow(() -> new WebServerException(ServerResponseEntity.fail("菜单不存在！")));
+        if (Objects.equals(sysMenu.getId(), sysMenu.getPid())) {
+            throw new WebServerException(ServerResponseEntity.fail("不能选择自己作为上级菜单！"));
+        }
+        if (Objects.nonNull(sysMenu.getPid())
+                && this.existsByIdAndPid(sysMenu.getId(), sysMenu.getPid())) {
+            throw new WebServerException(ServerResponseEntity.fail("不能选择自己的下级菜单作为上级菜单！"));
+        }
+        if (Objects.nonNull(sysMenu.getPid())
+                && !this.sysMenuDao.existsById(sysMenu.getPid())) {
+            throw new WebServerException(ServerResponseEntity.fail("上级菜单不存在！"));
+        }
+        if (!Objects.equals(oldSysMenu.getTitle(), sysMenu.getTitle())
+                && Objects.nonNull(this.sysMenuDao.findByTitle(sysMenu.getTitle()))) {
+            throw new WebServerException(ServerResponseEntity.fail("菜单标题已存在！"));
+        }
+        oldSysMenu.setPid(sysMenu.getPid());
+        BeanUtils.copyNonNullProperties(sysMenu, oldSysMenu);
+        this.sysMenuDao.save(oldSysMenu);
+    }
+
+    public void delete(Long id) {
+        SysMenu sysMenu = this.sysMenuDao.findById(id)
+                .orElseThrow(() -> new WebServerException(ServerResponseEntity.fail("菜单不存在！")));
+        if (this.sysMenuDao.existsByMenuId(sysMenu.getId())) {
+            throw new WebServerException(ServerResponseEntity.fail("此菜单已绑定角色，请先解绑角色下的菜单！"));
+        }
+        if (this.sysMenuDao.existsByPid(sysMenu.getId())) {
+            throw new WebServerException(ServerResponseEntity.fail("请先删除下级菜单！"));
+        }
         this.sysMenuDao.deleteById(id);
     }
 
-    public boolean existsByMenuId(Long menuId) {
-        return this.sysMenuDao.existsByMenuId(menuId);
-    }
-
-    public boolean existsByPid(Long pid) {
-        return this.sysMenuDao.existsByPid(pid);
-    }
-
-    public boolean existsById(Long id) {
-        return this.sysMenuDao.existsById(id);
-    }
-
-    public boolean existsByIdAndPid(Long id, Long pid) {
+    private boolean existsByIdAndPid(Long id, Long pid) {
         Set<Long> exists = new HashSet<>();
         SysMenu sysMenu = this.sysMenuDao.findById(pid).orElse(null);
         while (Objects.nonNull(sysMenu) && Objects.nonNull(sysMenu.getPid())) {
