+++ src/main/java/com/admin/web/service/SysRoleService.java
@@ -1,14 +1,17 @@
 package com.admin.web.service;
 
 import com.admin.web.dao.SysRoleDao;
+import com.admin.web.exception.WebServerException;
+import com.admin.web.model.ServerResponseEntity;
 import com.admin.web.model.SysRole;
 import com.admin.web.model.enums.Move;
+import com.admin.web.model.vo.MoveVo;
+import com.admin.web.utils.BeanUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 
 /**
  * @author znn
@@ -21,31 +24,14 @@ public class SysRoleService {
         this.sysRoleDao = sysRoleDao;
     }
 
-    public SysRole findByName(String name) {
-        return this.sysRoleDao.findByName(name);
-    }
-
-    public Optional<SysRole> findById(Long id) {
-        return this.sysRoleDao.findById(id);
-    }
-
-    public List<SysRole> findAll() {
+    public List<SysRole> all() {
         return this.sysRoleDao.findByOrderBySort();
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void save(SysRole sysRole) {
-        if (Objects.nonNull(sysRole.getId())) {
-            this.sysRoleDao.save(sysRole);
-        } else {
-            this.sysRoleDao.save(sysRole);
-            sysRole.setSort(sysRole.getId());
-            this.sysRoleDao.save(sysRole);
-        }
-    }
-
-    @Transactional(rollbackFor = Exception.class)
-    public void move(SysRole sysRole, Move move) {
+    public void move(MoveVo vo) {
+        SysRole sysRole = this.sysRoleDao.findById(vo.getId())
+                .orElseThrow(() -> new WebServerException(ServerResponseEntity.fail("角色不存在！")));
         List<SysRole> sysRoles = this.sysRoleDao.findByOrderBySort();
         int index = -1;
         for (int i = 0; i < sysRoles.size(); i++) {
@@ -54,8 +40,8 @@ public class SysRoleService {
                 break;
             }
         }
-        if (Move.UP == move ? index > 0 : index < sysRoles.size() - 1 && index != -1) {
-            SysRole oldSysRole = sysRoles.get(Move.UP == move ? index - 1 : index + 1);
+        if (Move.UP == vo.getMove() ? index > 0 : index < sysRoles.size() - 1 && index != -1) {
+            SysRole oldSysRole = sysRoles.get(Move.UP == vo.getMove() ? index - 1 : index + 1);
             Long oldSort = oldSysRole.getSort();
             oldSysRole.setSort(sysRole.getSort());
             sysRole.setSort(oldSort);
@@ -64,11 +50,33 @@ public class SysRoleService {
         }
     }
 
-    public void deleteById(Long id) {
-        this.sysRoleDao.deleteById(id);
+    @Transactional(rollbackFor = Exception.class)
+    public void create(SysRole sysRole) {
+        if (Objects.nonNull(this.sysRoleDao.findByName(sysRole.getName()))) {
+            throw new WebServerException(ServerResponseEntity.fail("角色名称已存在！"));
+        }
+        this.sysRoleDao.save(sysRole);
+        sysRole.setSort(sysRole.getId());
+        this.sysRoleDao.save(sysRole);
     }
 
-    public boolean existsByRoleId(Long roleId) {
-        return this.sysRoleDao.existsByRoleId(roleId);
+    public void update(SysRole sysRole) {
+        SysRole oldSysRole = this.sysRoleDao.findById(sysRole.getId())
+                .orElseThrow(() -> new WebServerException(ServerResponseEntity.fail("角色不存在！")));
+        if (!Objects.equals(oldSysRole.getName(), sysRole.getName())
+                && Objects.nonNull(this.sysRoleDao.findByName(sysRole.getName()))) {
+            throw new WebServerException(ServerResponseEntity.fail("角色名称已存在！"));
+        }
+        BeanUtils.copyNonNullProperties(sysRole, oldSysRole);
+        this.sysRoleDao.save(oldSysRole);
+    }
+
+    public void delete(Long id) {
+        SysRole sysRole = this.sysRoleDao.findById(id)
+                .orElseThrow(() -> new WebServerException(ServerResponseEntity.fail("角色不存在！")));
+        if (this.sysRoleDao.existsByRoleId(sysRole.getId())) {
+            throw new WebServerException(ServerResponseEntity.fail("此角色已绑定用户，请先解绑用户下的角色！"));
+        }
+        this.sysRoleDao.deleteById(id);
     }
 }
